+++ src/Formula.js
@@ -1,3 +1,4 @@
+import { forwardRef } from 'react';
 import RJSFForm from '@visma/rjsf-material-ui';
 import useForm from './useForm';
 import configToSchemas from './configToSchemas';
@@ -5,13 +6,13 @@ import { withFormulaProvider } from './Context';
 import withFormConfigLoader from './withFormConfigLoader';
 import useSubmit from './useSubmit';
 
-function Formula(props) {
+function Formula(props, ref) {
   const { config, ...otherProps } = props |> useSubmit;
   const schemaProps = config |> useForm |> configToSchemas;
 
-  return <RJSFForm {...otherProps} {...schemaProps} />;
+  return <RJSFForm ref={ref} {...otherProps} {...schemaProps} />;
 }
 
-export const Form = Formula |> withFormConfigLoader;
+export const Form = Formula |> forwardRef |> withFormConfigLoader;
 
 export default Form |> withFormulaProvider;